Project Executive Codes/requirements.txt
streamlit
google-generativeai
pandas
matplotlib
python-dotenv
